waapi/requirements.txt
Flask==2.0.1
cryptography==3.4.7
Flask-Cors==3.0.8
Flask-JWT-Extended==4.3.1
PyJWT==2.3.0
pymongo==4.0.0
dnspython==2.1.0
pyOpenSSL==20.0.1
python-dotenv==0.10.3
requests==2.26.0
gunicorn==20.1.0
PyYAML==5.4.1
six==1.16.0
XlsxWriter==3.0.1
pandas==1.3.3
openpyxl==3.0.9
redis==3.5.3
celery==5.2.3
typing==3.7.4.3
flower==1.0.0
Werkzeug==2.0.0
Bcrypt-Flask==1.0.2
eventlet==0.33.2
python-aes256==1.0.5
setuptools>=59.1.1
